refactoring: entity 이름 단수 -> 복수로 수정
