+++ src/modules/exams/exams.module.ts
@@ -2,8 +2,8 @@ import { Logger, Module } from '@nestjs/common';
 import { ExamsService } from './exams.service';
 import { ExamsController } from './exams.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Exams } from './entities/exam.entity';
-import { ExamUsers } from './entities/examusers.entity';
+import { Exams } from './entities/exams.entity';
+import { ExamUsers } from './entities/exams-users.entity';
 import { FilesModule } from '../files/files.module';
 
 @Module({
