+++ src/modules/users/dto/no-password-user.dto.ts
@@ -1,4 +1,4 @@
 import { OmitType } from '@nestjs/swagger';
-import { Users } from '../entities/user.entity';
+import { Users } from '../entities/users.entity';
 
 export class NoPasswordUserDto extends OmitType(Users, ['password']) {}
