+++ src/modules/users/dto/signup-user.dto.ts
@@ -1,5 +1,5 @@
 import { PickType } from '@nestjs/swagger';
-import { Users } from '../entities/user.entity';
+import { Users } from '../entities/users.entity';
 
 export class SignupUserDto extends PickType(Users, [
   'email',
