+++ src/modules/users/dto/update-user.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/swagger';
-import { Users } from '../entities/user.entity';
+import { Users } from '../entities/users.entity';
 
 export class UpdateUserDto extends PartialType(Users) {}
