+++ src/modules/users/entities/roles.entity.ts
@@ -1,7 +1,7 @@
 import { CommonEntity } from '../../../common/abstract/common.entity';
 import { Column, Entity, OneToMany } from 'typeorm';
 import { UserRole } from '../constants/user-role.enum';
-import { Users } from './user.entity';
+import { Users } from './users.entity';
 
 @Entity()
 export class Roles extends CommonEntity {
