+++ src/modules/users/users.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { Users } from './entities/user.entity';
-import { Roles } from './entities/role.entity';
+import { Users } from './entities/users.entity';
+import { Roles } from './entities/roles.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Users, Roles])],
